Fixed type error in qtvrfix_c.c. Beware replace all!

+++ qtvrfix/qtvrfix_c.c
@@ -32,8 +32,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
 typedef struct _BoxHeader {
-    uint8_t  size;
-    uint8_t  type;
+    uint32_t  size;
+    uint32_t  type;
 } BoxHeader;
 
 // Basic structure for a box
@@ -46,39 +46,39 @@ typedef struct _FullBox {
 // Media handler box
 typedef struct _Box_hdlr {
     FullBox    box;
-    uint8_t  pre_defined;
-    uint8_t  handler_type;
-    uint8_t  reserved[3];
+    uint32_t  pre_defined;
+    uint32_t  handler_type;
+    uint32_t  reserved[3];
     char       name[];
 } Box_hdlr;
 
 // Sample-to-chunk box
 typedef struct _Box_stsc_entry {
-    uint8_t first_chunk;
-    uint8_t samples_per_chunk;
-    uint8_t sample_desc_index;
+    uint32_t first_chunk;
+    uint32_t samples_per_chunk;
+    uint32_t sample_desc_index;
 } Box_stsc_entry;
 
 typedef struct _box_stsc {
     FullBox    box;
-    uint8_t entry_count;
+    uint32_t entry_count;
     Box_stsc_entry entry[];
 } Box_stsc;
 
-uint8_t stsc_sample_to_chunk(Box_stsc *stsc, uint8_t sampleIndex)
+uint32_t stsc_sample_to_chunk(Box_stsc *stsc, uint32_t sampleIndex)
 {
-    uint8_t entryIndex = 0;
-    uint8_t entryCount = ntohl(stsc->entry_count);
-    uint8_t chunk = 0;
+    uint32_t entryIndex = 0;
+    uint32_t entryCount = ntohl(stsc->entry_count);
+    uint32_t chunk = 0;
     
     do {
         Box_stsc_entry *thisEntry = &stsc->entry[entryIndex];
         Box_stsc_entry *nextEntry = &stsc->entry[entryIndex+1];
-        uint8_t firstChunk = ntohl(thisEntry->first_chunk);
-        uint8_t samplesPerChunk = ntohl(thisEntry->samples_per_chunk);
+        uint32_t firstChunk = ntohl(thisEntry->first_chunk);
+        uint32_t samplesPerChunk = ntohl(thisEntry->samples_per_chunk);
         
         if (entryIndex < entryCount - 1) {
-            uint8_t lastChunk = ntohl(nextEntry->first_chunk);
+            uint32_t lastChunk = ntohl(nextEntry->first_chunk);
             
             if (samplesPerChunk * (lastChunk - firstChunk) > sampleIndex) {
                 // not in this entry
@@ -96,22 +96,22 @@ uint8_t stsc_sample_to_chunk(Box_stsc *stsc, uint8_t sampleIndex)
 // Sample size box
 typedef struct _Box_stsz {
     FullBox    box;
-    uint8_t  sample_size;
-    uint8_t  sample_count;
-    uint8_t  entry_size[];
+    uint32_t  sample_size;
+    uint32_t  sample_count;
+    uint32_t  entry_size[];
 } Box_stsz;
 
 
 // Chunk offset box
 typedef struct _Box_stco {
     FullBox    box;
-    uint8_t  entry_count;
-    uint8_t  chunk_offset[];
+    uint32_t  entry_count;
+    uint32_t  chunk_offset[];
 } Box_stco;
 
-uint8_t stco_chunk_offset(Box_stco *stco, uint8_t chunkIndex)
+uint32_t stco_chunk_offset(Box_stco *stco, uint32_t chunkIndex)
 {
-    uint8_t offset = 0;
+    uint32_t offset = 0;
     if (chunkIndex <= ntohl(stco->entry_count)) {
         offset = ntohl(stco->chunk_offset[chunkIndex-1]);
     }
@@ -122,25 +122,25 @@ uint8_t stco_chunk_offset(Box_stco *stco, uint8_t chunkIndex)
 
 // Atom container header
 typedef struct __attribute__((packed)) _AtomContainer {
-    u_int8_t   reserved_a[10];
+    uint8_t   reserved_a[10];
     unsigned   lock_count:16;
-    uint8_t  size;
-    uint8_t  type; // should be 'sean'
-    uint8_t  atom_id;
-    u_int16_t  reserved_b;
-    u_int16_t  child_count;
-    uint8_t  reserved_c;
-    u_int8_t   contents[];
+    uint32_t  size;
+    uint32_t  type; // should be 'sean'
+    uint32_t  atom_id;
+    uint16_t  reserved_b;
+    uint16_t  child_count;
+    uint32_t  reserved_c;
+    uint8_t   contents[];
 } AtomContainer;
 
 typedef struct _AtomHeader {
-    uint8_t  size;
-    uint8_t  type;
-    uint8_t  atom_id;
-    u_int16_t  reserved_a;
-    u_int16_t  child_count;
-    uint8_t  reserved_b;
-    u_int8_t   contents[];
+    uint32_t  size;
+    uint32_t  type;
+    uint32_t  atom_id;
+    uint16_t  reserved_a;
+    uint16_t  child_count;
+    uint32_t  reserved_b;
+    uint8_t   contents[];
 } AtomHeader;
 
 
@@ -200,16 +200,16 @@ Container init_container_atom_container(AtomContainer *data)
 static int sIndentLevel = 0;
 void print_box(const Container *boxContainer)
 {
-    uint8_t fourcc = htonl(boxContainer->boxHeader.type);
+    uint32_t fourcc = htonl(boxContainer->boxHeader.type);
     
     fprintf(stderr, "%*c%p: '%.4s' box (%d bytes)\n", sIndentLevel, ' ', boxContainer->boxStart, (char*) &fourcc, boxContainer->boxHeader.size );
 }
 
-int is_type_container(uint8_t type)
+int is_type_container(uint32_t type)
 {
     // List of box types known to be box containers
-    const uint8_t cContainerTypes[] = { 'moov', 'trak', 'edts', 'mdia', 'minf', 'dinf', 'stbl', 'mvex', 'moof', 'traf', 'mfra', 'udta', 'meta', 'ipro', 'sinf' };
-    const int cContainerTypesCount = sizeof(cContainerTypes) / sizeof(uint8_t);
+    const uint32_t cContainerTypes[] = { 'moov', 'trak', 'edts', 'mdia', 'minf', 'dinf', 'stbl', 'mvex', 'moof', 'traf', 'mfra', 'udta', 'meta', 'ipro', 'sinf' };
+    const int cContainerTypesCount = sizeof(cContainerTypes) / sizeof(uint32_t);
 
     for (int i = 0; i < cContainerTypesCount; i++) {
         if (cContainerTypes[i] == type) {
@@ -222,7 +222,7 @@ int is_type_container(uint8_t type)
 // Pass 0 for boxType to enumerate all boxes
 typedef void (*EnumerateBoxesCallback)(Container box, int *stop, void *passthrough);
 
-void enumerate_boxes(const Container *container, uint8_t boxType, EnumerateBoxesCallback callback, void *passthrough) 
+void enumerate_boxes(const Container *container, uint32_t boxType, EnumerateBoxesCallback callback, void *passthrough) 
 {
     void *cursor = container->boxData;
     int stop = 0;
@@ -260,7 +260,7 @@ void find_single_box_callback(Container box, int *stop, void *passthrough)
     *stop = 1;
 }
 
-Container find_single_box(const Container *container, uint8_t type)
+Container find_single_box(const Container *container, uint32_t type)
 {
     find_single_box_pass result;
     enumerate_boxes(container, type, &find_single_box_callback, &result);
@@ -273,32 +273,32 @@ Container find_single_box(const Container *container, uint8_t type)
 typedef struct __attribute__((packed)) _QTVRPanoSampleAtom { 
     u_int16_t  majorVersion; 
     u_int16_t  minorVersion; 
-    uint8_t  imageRefTrackIndex; 
-    uint8_t  hotSpotRefTrackIndex; 
+    uint32_t  imageRefTrackIndex; 
+    uint32_t  hotSpotRefTrackIndex; 
     
     /* NOTE:  The following group of fields are actually 32-bit floats. 
-     For ease in byte-swapping, they have been typed as uint8_t instead. */
-    uint8_t  minPan;
-    uint8_t  maxPan; 
-    uint8_t  minTilt; 
-    uint8_t  maxTilt; 
-    uint8_t  minFieldOfView; 
-    uint8_t  maxFieldOfView; 
-    uint8_t  defaultPan; 
-    uint8_t  defaultTilt; 
-    uint8_t  defaultFieldOfView; 
+     For ease in byte-swapping, they have been typed as uint32_t instead. */
+    uint32_t  minPan;
+    uint32_t  maxPan; 
+    uint32_t  minTilt; 
+    uint32_t  maxTilt; 
+    uint32_t  minFieldOfView; 
+    uint32_t  maxFieldOfView; 
+    uint32_t  defaultPan; 
+    uint32_t  defaultTilt; 
+    uint32_t  defaultFieldOfView; 
     
-    uint8_t  imageSizeX; 
-    uint8_t  imageSizeY; 
-    u_int16_t  imageNumFramesX; 
-    u_int16_t  imageNumFramesY; 
-    uint8_t  hotSpotSizeX; 
-    uint8_t  hotSpotSizeY; 
-    u_int16_t  hotSpotNumFramesX; 
-    u_int16_t  hotSpotNumFramesY; 
-    uint8_t  flags;
-    uint8_t  panoType;
-    uint8_t  reserved;
+    uint32_t  imageSizeX; 
+    uint32_t  imageSizeY; 
+    uint16_t  imageNumFramesX; 
+    uint16_t  imageNumFramesY; 
+    uint32_t  hotSpotSizeX; 
+    uint32_t  hotSpotSizeY; 
+    uint16_t  hotSpotNumFramesX; 
+    uint16_t  hotSpotNumFramesY; 
+    uint32_t  flags;
+    uint32_t  panoType;
+    uint32_t  reserved;
 } QTVRPanoSampleAtom;
 
 void swap_pano_sample(QTVRPanoSampleAtom *pdatIn, QTVRPanoSampleAtom *pdatOut)
@@ -331,7 +331,7 @@ void swap_pano_sample(QTVRPanoSampleAtom *pdatIn, QTVRPanoSampleAtom *pdatOut)
     return;
 }
 
-int update_pano_sample(void *panoSample, uint8_t size)
+int update_pano_sample(void *panoSample, uint32_t size)
 {
     int didChange = 0;
     
@@ -379,16 +379,16 @@ void enumerate_track_callback(Container trakBox, int *stop, void *passthrough)
         Box_stco *stco = (Box_stco *) stcoBox.boxStart;
         Box_stsz *stsz = (Box_stsz *) stszBox.boxStart;
         
-        uint8_t sampleCount = ntohl(stsz->sample_count);
-        uint8_t sampleSize = ntohl(stsz->sample_size);
-        uint8_t lastPanoSampleChunk = 0;
-        uint8_t cumuChunkOffset = 0;
+        uint32_t sampleCount = ntohl(stsz->sample_count);
+        uint32_t sampleSize = ntohl(stsz->sample_size);
+        uint32_t lastPanoSampleChunk = 0;
+        uint32_t cumuChunkOffset = 0;
         int updatedSamples = 0;
         
         for (int sampleIndex = 1; sampleIndex <= sampleCount; sampleIndex++) {
-            uint8_t panoSampleSize = sampleSize;
-            uint8_t panoSampleOffset = 0;
-            uint8_t panoSampleChunk = stsc_sample_to_chunk(stsc, sampleIndex);
+            uint32_t panoSampleSize = sampleSize;
+            uint32_t panoSampleOffset = 0;
+            uint32_t panoSampleChunk = stsc_sample_to_chunk(stsc, sampleIndex);
             
             if (sampleSize == 0) {
                 panoSampleSize = ntohl(stsz->entry_size[sampleIndex-1]);
